# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "EntriesBody": {
        "properties": {
          "distance_m": {
            "exclusiveMinimum": 0.0,
            "title": "Distance M",
            "type": "integer"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/EntryItem"
            },
            "title": "Entries",
            "type": "array"
          },
          "field_size": {
            "exclusiveMinimum": 0.0,
            "title": "Field Size",
            "type": "integer"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "jyo_cd": {
            "maxLength": 2,
            "minLength": 2,
            "title": "Jyo Cd",
            "type": "string"
          },
          "race_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Race Class"
          },
          "race_date": {
            "format": "date",
            "title": "Race Date",
            "type": "string"
          },
          "race_key": {
            "pattern": "^\\d{16}$",
            "title": "Race Key",
            "type": "string"
          },
          "track_condition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Track Condition"
          },
          "track_type": {
            "title": "Track Type",
            "type": "string"
          },
          "weather": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weather"
          }
        },
        "required": [
          "race_key",
          "race_date",
          "jyo_cd",
          "distance_m",
          "track_type",
          "field_size",
          "entries"
        ],
        "title": "EntriesBody",
        "type": "object"
      },
      "EntryDetailSchema": {
        "description": "出走馬の詳細（確定済みなら成績を含む）。",
        "properties": {
          "finish_pos": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finish Pos"
          },
          "frame_no": {
            "title": "Frame No",
            "type": "integer"
          },
          "horse_no": {
            "title": "Horse No",
            "type": "integer"
          },
          "ketto_num": {
            "title": "Ketto Num",
            "type": "string"
          },
          "pci_actual": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pci Actual"
          },
          "running_style": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Running Style"
          }
        },
        "required": [
          "horse_no",
          "frame_no",
          "ketto_num"
        ],
        "title": "EntryDetailSchema",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HorseBody": {
        "properties": {
          "birth_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Year"
          },
          "ketto_num": {
            "maxLength": 10,
            "minLength": 10,
            "title": "Ketto Num",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sex"
          }
        },
        "required": [
          "ketto_num",
          "name"
        ],
        "title": "HorseBody",
        "type": "object"
      },
...
      "IngestLogBody": {
        "description": "バッチ実行ログの記録リクエスト。",
        "properties": {
          "batch_date": {
            "format": "date",
            "title": "Batch Date",
            "type": "string"
          },
          "error_msg": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Msg"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "step": {
            "title": "Step",
            "type": "string"
          }
        },
        "required": [
          "batch_date",
          "step",
          "mode",
          "started_at"
        ],
        "title": "IngestLogBody",
        "type": "object"
      },
      "IngestLogResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id"
        ],
        "title": "IngestLogResponse",
        "type": "object"
      },
      "IngestResponse": {
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "integer"
          },
          "message": {
            "default": "ok",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "accepted"
        ],
        "title": "IngestResponse",
        "type": "object"
      },
...
      "JockeyBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "code",
          "name"
        ],
        "title": "JockeyBody",
        "type": "object"
      },
...
      "RaceDetailSchema": {
        "description": "レース詳細（core 層の情報 + 確定指標）。",
        "properties": {
          "distance_m": {
            "title": "Distance M",
            "type": "integer"
          },
          "entries": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EntryDetailSchema"
            },
            "title": "Entries",
            "type": "array"
          },
          "field_size": {
            "title": "Field Size",
            "type": "integer"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "jyo_cd": {
            "title": "Jyo Cd",
            "type": "string"
          },
          "pci3_actual": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pci3 Actual"
          },
          "race_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Race Class"
          },
          "race_date": {
            "title": "Race Date",
            "type": "string"
          },
          "race_key": {
            "title": "Race Key",
            "type": "string"
          },
          "rpci_actual": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rpci Actual"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "track_condition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Track Condition"
          },
          "track_type": {
            "title": "Track Type",
            "type": "string"
          },
          "weather": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weather"
          }
        },
        "required": [
          "race_key",
          "race_date",
          "jyo_cd",
          "distance_m",
          "track_type",
          "status",
          "field_size"
        ],
        "title": "RaceDetailSchema",
        "type": "object"
      },
...
      "ResultBody": {
        "properties": {
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "race_key": {
            "pattern": "^\\d{16}$",
            "title": "Race Key",
            "type": "string"
          },
          "race_l3f": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Race L3F"
          },
          "race_s3f": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Race S3F"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/ResultItem"
            },
            "title": "Results",
            "type": "array"
          },
          "track_condition": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Track Condition"
          },
          "weather": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weather"
          }
        },
        "required": [
          "race_key",
          "results"
        ],
        "title": "ResultBody",
        "type": "object"
      },
...
      "ResultResponse": {
        "properties": {
          "entry_pcis": {
            "additionalProperties": {
              "type": "number"
            },
            "default": {},
            "title": "Entry Pcis",
            "type": "object"
          },
          "formula_version": {
            "title": "Formula Version",
            "type": "string"
          },
          "pci3": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pci3"
          },
          "race_key": {
            "title": "Race Key",
            "type": "string"
          },
          "rpci": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rpci"
          }
        },
        "required": [
          "race_key",
          "rpci",
          "pci3",
          "formula_version"
        ],
        "title": "ResultResponse",
        "type": "object"
      },
...
      "TrainerBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "code",
          "name"
        ],
        "title": "TrainerBody",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "競馬展開予想 SaaS — 想定RPCI・PAI・展開シナリオを提供する REST API",
    "title": "PCI App API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/races/{race_key}": {
      "get": {
        "description": "レースの基本情報・出走馬・確定指標を返す。",
        "operationId": "get_race_detail_api_v1_races__race_key__get",
        "parameters": [
          {
            "description": "16桁のレースキー",
            "in": "path",
            "name": "race_key",
            "required": true,
            "schema": {
              "description": "16桁のレースキー",
              "pattern": "^\\d{16}$",
              "title": "Race Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RaceDetailSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Race Detail",
        "tags": [
          "races"
        ]
      }
    },
...
    "/internal/ingest/entries": {
      "post": {
        "description": "出走表登録（RegisterRaceEntriesUseCase を呼び出す）。",
        "operationId": "ingest_entries_internal_ingest_entries_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-ingest-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ingest-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntriesBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Entries",
        "tags": [
          "ingest"
        ]
      }
    },
    "/internal/ingest/horses": {
      "post": {
        "description": "馬マスタ一括 Upsert。",
        "operationId": "ingest_horses_internal_ingest_horses_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-ingest-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ingest-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/HorseBody"
                },
                "title": "Body",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Horses",
        "tags": [
          "ingest"
        ]
      }
    },
    "/internal/ingest/jockeys": {
      "post": {
        "description": "騎手マスタ一括 Upsert。",
        "operationId": "ingest_jockeys_internal_ingest_jockeys_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-ingest-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ingest-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/JockeyBody"
                },
                "title": "Body",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Jockeys",
        "tags": [
          "ingest"
        ]
      }
    },
    "/internal/ingest/log": {
      "post": {
        "description": "バッチ取り込みの実行ログを記録する（監査証跡・再実行判定用）。",
        "operationId": "write_ingest_log_internal_ingest_log_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-ingest-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ingest-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestLogBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestLogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Write Ingest Log",
        "tags": [
          "ingest"
        ]
      }
    },
    "/internal/ingest/races/{race_key}": {
      "delete": {
        "description": "取り込み対象外になったレースを、関連する予想データごと削除する。",
        "operationId": "delete_ingested_race_internal_ingest_races__race_key__delete",
        "parameters": [
          {
            "in": "path",
            "name": "race_key",
            "required": true,
            "schema": {
              "title": "Race Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-ingest-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ingest-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Ingested Race",
        "tags": [
          "ingest"
        ]
      }
    },
    "/internal/ingest/results": {
      "post": {
        "description": "確定成績登録（RecordRaceResultUseCase を呼び出す）。",
        "operationId": "ingest_results_internal_ingest_results_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-ingest-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ingest-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResultBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Results",
        "tags": [
          "ingest"
        ]
      }
    },
    "/internal/ingest/trainers": {
      "post": {
        "description": "調教師マスタ一括 Upsert。",
        "operationId": "ingest_trainers_internal_ingest_trainers_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-ingest-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Ingest-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/TrainerBody"
                },
                "title": "Body",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Trainers",
        "tags": [
          "ingest"
        ]
      }
    }
  }
}